Update to remove unnecessary errors, update documentation and a small api simplification
